+++ javascript/src/deckgl-widget.js
@@ -1,5 +1,6 @@
 import "../styles/default.css";
 
+import Viz from "./viz";
 import {
   createControlGroups,
   createTooltip,
@@ -10,7 +11,6 @@ import {
   logVersions,
   fixLayerProperties,
   convertColor } from "./utils";
-import parseLayerProps from "./layer";
 import addJSONEditor from "./json-editor";
 
 if (!global._deckWidget) {
@@ -71,43 +71,4 @@ const parseSources = (sources) => sources.map(({ id, data, df }) => {
   if (df) data = HTMLWidgets.dataframeToD3(data);
 
   return { id, data };
-});
-
-// TODO: Move to separte file, is the widgetElement really needed here?
-const Viz = ({ deckGL, layerDefs, sources, widgetElement }) => ({
-  deckGL,
-  layerDefs,
-  sources,
-  widgetElement,
-
-  _getContainer() {
-    return this.deckGL.props.container;
-  },
-
-  setLayerDefs(layerDefs) {
-    this.layerDefs = layerDefs;
-  },
-
-  render() {
-    const layers = this.layerDefs.map(layerDef => {
-      /*
-      if (layerDef.properties.dataframeToD3) {
-        layerDef.data = HTMLWidgets.dataframeToD3(layerDef.data);
-        layerDef.properties.dataframeToD3 = false;
-      }
-      */
-      /*
-      if (layerDef.source.df) {
-        layerDef.source.data = HTMLWidgets.dataframeToD3(layerDef.source.data);
-        layerDef.source.df = false;
-      }
-      */
-
-      layerDef.properties.data = this.sources.filter(source =>
-        source.id === layerDef.source)[0].data; //layerDef.source.data;
-      const props = parseLayerProps(layerDef.properties, this.widgetElement);
-      return new deck[layerDef.className](props);
-    });
-    this.deckGL.setProps({ layers: layers });
-  }
 });
